+++ src/PaymentGateway.Application/Models/AcquiringBankPaymentRequest.cs
@@ -5,5 +5,5 @@ public record AcquiringBankPaymentRequest(
     int ExpiryMonth,
     int ExpiryYear,
     string Currency,
-    long Amount,
+    int Amount,
     string Cvv);
